+++ common/RELEASES.md
@@ -631,3 +631,13 @@ Based on:
 - Speakeasy CLI 1.59.0 (2.65.0) https://github.com/speakeasy-api/speakeasy
 ### Releases
 - [NPM v0.35.1] https://www.npmjs.com/package/@codat/common/v/0.35.1 - common
+
+## 2023-08-17 00:03:42
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Common.yaml
+- Speakeasy CLI 1.73.1 (2.84.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.36.0] common
+### Releases
+- [NPM v0.36.0] https://www.npmjs.com/package/@codat/common/v/0.36.0 - common